# perf(ext/http): lazy load headers