Extracted from #15032 

No headers:

```
# main
Running 10s test @ http://127.0.0.1:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    95.89us   59.87us   4.87ms   97.74%
    Req/Sec    49.72k     3.66k   53.86k    94.06%
  Latency Distribution
     50%   92.00us
     75%  106.00us
     90%  122.00us
     99%  189.00us
  999544 requests in 10.10s, 104.86MB read
Requests/sec:  98968.27
Transfer/sec:     10.38MB
```

```
# This PR
Running 10s test @ http://127.0.0.1:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    90.65us   30.14us   1.26ms   86.37%
    Req/Sec    51.76k     2.64k   54.26k    99.01%
  Latency Distribution
     50%   87.00us
     75%  102.00us
     90%  116.00us
     99%  161.00us
  1040014 requests in 10.10s, 109.10MB read
Requests/sec: 102975.04
Transfer/sec:     10.80MB
```

With headers (`cli/bench/http/deno_http_read_headers.js`):

```
# main
Running 20s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   115.21us   44.89us   5.24ms   90.97%
    Req/Sec    40.86k     1.88k   42.04k    98.01%
  Latency Distribution
     50%  113.00us
     75%  130.00us
     90%  148.00us
     99%  190.00us
  1634863 requests in 20.10s, 219.84MB read
Requests/sec:  81337.68
Transfer/sec:     10.94MB
```

```
# This patch
Running 20s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   114.45us   32.51us   2.51ms   80.05%
    Req/Sec    40.99k     1.06k   42.27k    95.52%
  Latency Distribution
     50%  112.00us
     75%  130.00us
     90%  148.00us
     99%  184.00us
  1639531 requests in 20.10s, 220.46MB read
Requests/sec:  81576.20
Transfer/sec:     10.97MB
```

With `hey`:

```
# This patch
hey -n 1000000  http://127.0.0.1:4500

Summary:
  Total:        11.1579 secs
  Slowest:      0.0263 secs
  Fastest:      0.0000 secs
  Average:      0.0006 secs
  Requests/sec: 89622.9349
```

```
# Node v18.0.0
hey -n 1000000  http://127.0.0.1:4500

Summary:
  Total:        11.9243 secs
  Slowest:      0.0196 secs
  Fastest:      0.0000 secs
  Average:      0.0006 secs
  Requests/sec: 83862.4296
```